refactor: index 파일 명 오류 수정하고 import 방식 수정
